requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
starlette
pydantic
